+++ www/static/biquge/tempbookcase.js
@@ -9,55 +9,81 @@ LastRead.prototype = {
     if (!(bid && url && bookname && chaptername && author && readtime)) return;
     var v = bid + "#" + url + "#" + bookname + "#" + chaptername + "#" + author + "#" + readtime;
 
-    var aBooks = lastread.getBook();
+    var aBooks = this.getBook();
     var aBid = [];
     for (var i = 0; i < aBooks.length; i++) aBid.push(aBooks[i][0]);
 
-    if ($.inArray(bid, aBid) != -1) {
-      lastread.remove(bid);
+    if ($.inArray(String(bid), aBid) !== -1) {
+      this.remove(bid);
     } else {
       while (aBooks.length >= bookmax) {
-        lastread.remove(aBooks[0][0]);
-        aBooks = lastread.getBook();
+        this.remove(aBooks[0][0]);
+        aBooks = this.getBook();
       }
     }
     this.setItem(bid, v);
     this.setBook(bid);
   },
   get: function (k) {
-    return this.getItem(k) ? this.getItem(k).split("#") : "";
+    var value = this.getItem(k);
+    return value ? value.split("#") : [];
   },
   remove: function (k) {
     this.removeItem(k);
     this.removeBook(k);
   },
   setBook: function (v) {
-    var reg = new RegExp("(^|#)" + v);
-    var books = this.getBook();
-    if (!reg.test(books)) books.push(v);
-    this.setItem(this.bookList, books.join("#"));
+    var ids = this.getBookIds();
+    var next = [];
+    v = String(v);
+    for (var i = 0; i < ids.length; i++) {
+      if (ids[i] && ids[i] !== v) next.push(ids[i]);
+    }
+    next.push(v);
+    this.setItem(this.bookList, next.join("#"));
+  },
+  getBookIds: function () {
+    var raw = this.getItem(this.bookList);
+    if (!raw) return [];
+    var ids = raw.split("#");
+    var next = [];
+    for (var i = 0; i < ids.length; i++) {
+      var id = String(ids[i] || "").trim();
+      if (id) next.push(id);
+    }
+    return next;
   },
   getBook: function () {
-    var bookList = this.getItem(this.bookList);
-    return bookList ? bookList.split("#").map(this.get, this) : [];
-  },
-  removeBook: function (v) {
-    var books = this.getBook();
-    for (var i = 0; i < books.length; i++) {
-      if (books[i][0] == v) {
-        books.splice(i, 1);
-        i--;
+    var ids = this.getBookIds();
+    var books = [];
+    var validIds = [];
+    for (var i = 0; i < ids.length; i++) {
+      var tem = this.get(ids[i]);
+      if (tem.length >= 4 && tem[0] && tem[1] && tem[2]) {
+        books.push(tem);
+        validIds.push(String(tem[0]));
       }
     }
-    var str = [];
-    for (var j = 0; j < books.length; j++) str.push(books[j][0]);
-    this.setItem(this.bookList, str.join("#"));
+    var normalized = validIds.join("#");
+    if (normalized !== ids.join("#")) {
+      this.setItem(this.bookList, normalized);
+    }
+    return books;
+  },
+  removeBook: function (v) {
+    var ids = this.getBookIds();
+    var next = [];
+    v = String(v);
+    for (var i = 0; i < ids.length; i++) {
+      if (ids[i] !== v) next.push(ids[i]);
+    }
+    this.setItem(this.bookList, next.join("#"));
   },
   setItem: function (k, v) {
     localStorage.setItem(k, v);
   },
   getItem: function (k) {
-    return localStorage.getItem(k);
+    return localStorage.getItem(k) || "";
   },
   removeItem: function (k) {
     localStorage.removeItem(k);
@@ -93,6 +119,15 @@ function fmtReadTime(t) {
   return t;
 }
 
+function escapeHtml(value) {
+  return String(value || "")
+    .replace(/&/g, "&amp;")
+    .replace(/</g, "&lt;")
+    .replace(/>/g, "&gt;")
+    .replace(/"/g, "&quot;")
+    .replace(/'/g, "&#39;");
+}
+
 function removebook(bookid) {
   lastread.remove(bookid);
   showtempbooks();
@@ -101,22 +136,26 @@ function removebook(bookid) {
 function showtempbooks() {
   var books = lastread.getBook().reverse();
   var bookhtml = "";
+  var rendered = 0;
   if (books.length) {
     for (var i = 0; i < books.length; i++) {
-      if (i < bookmax) {
-        var rt = fmtReadTime(books[i][5]);
-        bookhtml += '<div class="recentread-main"><a href="' + books[i][1] + '">';
-        bookhtml += "<span>" + (i + 1) + "</span>";
-        bookhtml += "<span>" + (books[i][2] || "") + "</span>";
-        bookhtml += "<span>" + (books[i][3] || "") + "</span>";
-        bookhtml += "<span>" + (books[i][4] || "") + "</span>";
-        bookhtml += "<span>" + rt + "</span>";
-        bookhtml += "</a>";
-        bookhtml += '<a href="javascript:removebook(\'' + books[i][0] + '\')" onclick="return confirm(\'确定要将本书移除吗？\')">移除</a></div>';
-      }
+      if (rendered >= bookmax) break;
+      var item = books[i];
+      if (!item || item.length < 4 || !item[0] || !item[1] || !item[2]) continue;
+      var rt = fmtReadTime(item[5]);
+      rendered++;
+      bookhtml += '<div class="recentread-main"><a href="' + escapeHtml(item[1]) + '">';
+      bookhtml += "<span>" + rendered + "</span>";
+      bookhtml += "<span>" + escapeHtml(item[2]) + "</span>";
+      bookhtml += "<span>" + escapeHtml(item[3] || "") + "</span>";
+      bookhtml += "<span>" + escapeHtml(item[4] || "") + "</span>";
+      bookhtml += "<span>" + escapeHtml(rt) + "</span>";
+      bookhtml += "</a>";
+      bookhtml += '<a href="javascript:removebook(\'' + escapeHtml(item[0]) + '\')" onclick="return confirm(\'确定要将本书移除吗？\')">移除</a></div>';
     }
-  } else {
-    bookhtml += "<span>没有阅读记录。</span>";
+  }
+  if (!bookhtml) {
+    bookhtml = '<span style="display:block;padding:12px 10px;color:#888;">没有阅读记录。</span>';
   }
   $("#tempBookcase").html(bookhtml);
 }